requirements.txt
colorama==0.4.6
contourpy==1.3.3
cvxopt==1.3.2
cycler==0.12.1
deprecation==2.1.0
filelock==3.13.1
fonttools==4.59.0
fsspec==2024.6.1
graphviz==0.21
intervaltree==3.1.0
Jinja2==3.1.4
joblib==1.5.1
kiwisolver==1.4.8
lxml==6.0.0
MarkupSafe==2.1.5
matplotlib==3.10.5
mpmath==1.3.0
networkx==3.3
numpy==2.1.2
packaging==25.0
pandas==2.3.1
pillow==11.0.0
pm4py==2.7.16
pydotplus==2.0.2
pyparsing==3.2.3
python-dateutil==2.9.0.post0
pytz==2025.2
scikit-learn==1.7.1
scipy==1.16.1
seaborn==0.13.2
six==1.17.0
sortedcontainers==2.4.0
sympy==1.13.3
threadpoolctl==3.6.0
torch==2.7.1+cu128
torchaudio==2.7.1+cu128
torchvision==0.22.1+cu128
tqdm==4.67.1
typing_extensions==4.12.2
tzdata==2025.2
